Use yugabyte-bash-common for venv management in yugabyte-db
Summary:
This replaces the two custom create_venv implementations with the one from yugabyte-bash-common.

The yugabyte-bash-common code allows specifying a specific python version to use and also has
builtin logic to ensure the virtual env's integrity.  As part of this both requirements.txt files
were refrozen with python 3.8.  Once this is in and we are sure it is working, we can bump it up
to something newer.

Test Plan: Jenkins: compile only

Reviewers: devops

Subscribers: devops, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39281